@@ -21,7 +21,7 @@
     </modules>
 
     <properties>
-            <spring-boot.version>2.7.8</spring-boot.version>
+            <spring-boot.version>3.0.2</spring-boot.version>
     </properties>
 
     <dependencyManagement>
