@@ -2,11 +2,11 @@ package de.niko.pcstore.configuration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import de.niko.pcstore.dto.ErrorDTO;
+import jakarta.servlet.http.HttpServletResponse;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import javax.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
